requirements.txt
pandas
pdfplumber
matplotlib
google-generativeai